backend/requirements.txt
rasa==3.6.0
rasa-sdk==3.6.0
PyYAML==5.3.1
numpy==1.23.5
packaging==20.9
aio-pika==8.2.3
pytesseract==0.3.8
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.4
pillow==10.4.0
Cython==0.29.36
pip>=25.1.1
setuptools<81